+++ src/functions/VectorNorm.java
@@ -2,14 +2,14 @@ package functions;
 
 public class VectorNorm {
 
-	float VectorNrm(float[] a) {	//Norma de um vetor
+	double VectorNrm(double[] a) {	//Norma de um vetor
 		if(a.length != 3) {
 			System.out.println("Vetor mal definido");
 			return 0;
 		} else {
-			float result = 0;
+			double result = 0;
 			
-			result = (float) Math.sqrt(a[0] * a[0] + a[1]  * a[1] + a[2] * a[2]);
+			result = (double) Math.sqrt(a[0] * a[0] + a[1]  * a[1] + a[2] * a[2]);
 			
 			return result;
 		}
